Conflicts:
	python/yt/wrapper/tests/test_native_mode.py
	scripts/reshard.py
	yt/server/cell_master/bootstrap.cpp
	yt/server/data_node/block_store.cpp
	yt/server/data_node/chunk_cache.cpp
	yt/server/exec_agent/job.cpp
	yt/server/exec_agent/slot.cpp
	yt/server/exec_agent/slot.h
	yt/server/exec_agent/slot_manager.h
	yt/server/job_proxy/user_job.cpp
	yt/server/scheduler/map_controller.cpp
	yt/server/scheduler/merge_controller.cpp
	yt/server/scheduler/operation_controller_detail.cpp
	yt/ytlib/api/client.h
	yt/ytlib/chunk_client/replication_reader.cpp
	yt/ytlib/driver/table_commands.h
	yt/ytlib/query_client/plan_fragment.proto
	yt/ytlib/table_client/pipe.cpp

+++ yt/core/compression/lz.h
@@ -7,7 +7,7 @@ namespace NCompression {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-int Lz4CompressionBound(const std::vector<int>& lengths);
+size_t Lz4CompressionBound(const std::vector<int>& lengths);
 
 void Lz4Compress(bool highCompression, StreamSource* source, TBlob* output);
